+++ protaskinate/routes/project.py
@@ -12,9 +12,11 @@ from wtforms.validators import DataRequired, Optional
 from protaskinate.entities.task import TaskPriority, TaskStatus
 from protaskinate.services import (comment_service, project_service,
                                    task_service, user_service)
+from protaskinate.utils.project import (project_read_access_required,
+                                        project_update_access_required,
+                                        task_update_access_required)
 
 blueprint = Blueprint("project", __name__)
-
 
 class CreateTaskForm(FlaskForm):
     """Form for creating a task"""
@@ -46,11 +48,12 @@ class CreateCommentForm(FlaskForm):
 @login_required
 def project_list_route():
     """Render the projects page"""
-    projects = project_service.get_all()
-    return render_template("project_list.html", projects=projects)
+    projects_with_roles = project_service.get_all_by_user_with_role(current_user.id)
+    return render_template("project_list.html", projects_with_roles=projects_with_roles)
 
 @blueprint.route("/projects/<int:project_id>", methods=["GET", "POST"])
 @login_required
+@project_read_access_required
 def project_view_route(project_id: int):
     """Render the single project view page"""
     form = CreateTaskForm(request.form)
@@ -61,32 +64,41 @@ def project_view_route(project_id: int):
         assignee_id = form.assignee_id.data if form.assignee_id.data != 0 else None
         deadline = form.deadline.data.isoformat() if form.deadline.data else None
         description = form.description.data if form.description.data else None
-        task_service.create(title=title,
-                            status=status,
-                            priority=priority,
-                            creator_id=current_user.id,
-                            created_at=datetime.now().isoformat(),
-                            updated_at=datetime.now().isoformat(),
-                            assignee_id=assignee_id,
-                            deadline=deadline,
-                            project_id=project_id,
-                            description=description)
-        return redirect(request.url)
+        if not project_service.check_user_write_access(current_user.id, project_id):
+            flash("You do not have write-access to this project", "error")
+        else:
+            task_service.create(title=title,
+                                status=status,
+                                priority=priority,
+                                creator_id=current_user.id,
+                                created_at=datetime.now().isoformat(),
+                                updated_at=datetime.now().isoformat(),
+                                assignee_id=assignee_id,
+                                deadline=deadline,
+                                project_id=project_id,
+                                description=description)
+            form.title.data = ""
+            form.assignee_id.data = 0
+            form.deadline.data = None
+            form.description.data = ""
 
     project = project_service.get_by_id(project_id)
     tasks = task_service.get_all_by_project(project_id,
                                             order_by_fields=["priority", "created_at"],
                                             reverse=[True, False])
+    user_project_role = project_service.get_user_role(current_user.id, project_id)
     users_dict = {user.id: user for user in user_service.get_all()}
 
     return render_template("project_view.html",
                            form=form,
                            project=project,
+                           user_project_role=user_project_role,
                            tasks=tasks,
                            users_dict=users_dict)
 
 @blueprint.route("/projects/<int:project_id>/delete", methods=["POST"])
 @login_required
+@project_update_access_required
 def project_delete_route(project_id: int):
     """Delete a project"""
     project_service.delete(project_id)
@@ -94,6 +106,7 @@ def project_delete_route(project_id: int):
 
 @blueprint.route("/projects/<int:project_id>/tasks/<int:task_id>", methods=["GET", "POST"])
 @login_required
+@project_read_access_required
 def project_task_view_route(project_id: int, task_id: int):
     """View of a single task in a project"""
     form = CreateCommentForm(request.form)
@@ -109,6 +122,7 @@ def project_task_view_route(project_id: int, task_id: int):
     task = task_service.get_by_id_and_project(task_id, project_id)
     comments = comment_service.get_all_by_task(task_id)
     project = project_service.get_by_id(project_id)
+    user_project_role = project_service.get_user_role(current_user.id, project_id)
     users_dict = {user.id: user for user in user_service.get_all()}
 
     if not project:
@@ -117,10 +131,12 @@ def project_task_view_route(project_id: int, task_id: int):
         return redirect(url_for("project.project_view_route", project_id=project_id))
 
     return render_template("project_task_view.html", project=project,
-                           task=task, comments=comments, users_dict=users_dict, form=form)
+                           task=task, comments=comments, users_dict=users_dict,
+                           user_project_role=user_project_role, form=form)
 
 @blueprint.route("/projects/<int:project_id>/tasks/<int:task_id>/edit", methods=["POST"])
 @login_required
+@task_update_access_required
 def project_task_edit_route(project_id: int, task_id: int):
     """Edit a task in a project"""
     data = request.form
@@ -143,6 +159,7 @@ def project_task_edit_route(project_id: int, task_id: int):
 
 @blueprint.route("/projects/<int:project_id>/tasks/<int:task_id>/delete", methods=["POST"])
 @login_required
+@task_update_access_required
 def project_task_delete_route(project_id: int, task_id: int):
     """Delete a task from a project"""
     task_service.delete(task_id, project_id)