+++ protaskinate/services/user_service.py
@@ -1,5 +1,4 @@
 """protaskinate/services/user_service.py"""
-
 from typing import List, Optional
 
 from werkzeug.security import generate_password_hash
@@ -10,11 +9,6 @@ from protaskinate.repositories import user_repository
 
 class UserService:
     """Class representing a service for users"""
-    def login(self, username: str, password: str) -> Optional[User]:
-        """Login a user"""
-        if user_repository.verify_password(username, password):
-            return user_repository.get({"username": username})
-        return None
 
     def get_all(self) -> List[User]:
         """Get all users"""
@@ -27,6 +21,12 @@ class UserService:
     def get_by_username(self, username: str) -> Optional[User]:
         """Get a user by username"""
         return user_repository.get({"username": username})
+
+    def login(self, username: str, password: str) -> Optional[User]:
+        """Login a user"""
+        if user_repository.verify_password(username, password):
+            return user_repository.get({"username": username})
+        return None
 
     def register(self, username: str, password: str) -> Optional[User]:
         """Register a user"""